feat: Add authentication, add dockerfile and docker compose and documentation

+++ carford_car_shop/blueprints/car_bp.py
@@ -1,8 +1,10 @@
 from flask_openapi3 import APIBlueprint, Tag
+from flask import request
 
+from logger import logger
 from models import Car, Customer, ModelEnum, ColorEnum, Session
 from schemas import *
-
+from utils import jwt_decode
 
 tag = Tag(name="Cars API", description="Routes to insert, read and remove cars.")
 
@@ -13,6 +15,8 @@ car_bp = APIBlueprint(
     abp_tags=[tag],
     doc_ui=True,
 )
+
+security = [{"jwt": []}]
 
 
 @car_bp.get("/cars", responses={"200": CarResponseSchema})
@@ -55,7 +59,7 @@ def get_car_by_id(path: IdPathSchema):
     return {"message": f"Car of ID {car_id} not found."}, 404
 
 
-@car_bp.post("/cars", responses={"200": MessageSchema})
+@car_bp.post("/cars", responses={"200": MessageSchema}, security=security)
 def add_car(body: CarBodySchema):
     """Add a new car to the database
 
@@ -64,6 +68,13 @@ def add_car(body: CarBodySchema):
     color = 'yellow', 'blue' or 'gray'
     owner_id = <customer.id>
     """
+    token = str(request.authorization).split(" ")[-1]
+    token_check = jwt_decode(token)
+
+    if not token_check:
+        error_msg = "Login required for insert operation"
+        return {"error": error_msg}, 403
+
     session = Session()
 
     if not hasattr(ModelEnum, body.model) or not hasattr(ColorEnum, body.color):
@@ -71,7 +82,6 @@ def add_car(body: CarBodySchema):
         return {"error": error_msg}, 400
 
     customer = session.query(Customer).filter(Customer.id == body.owner_id).first()
-
     if not customer.can_add_new_car():
         return {"message": "Customer own 3 cars and can not add a new one."}, 400
 
@@ -80,15 +90,23 @@ def add_car(body: CarBodySchema):
         session.add(car)
         session.commit()
     except Exception as err:
-        print(err)
+        error_msg = f"error msg: {err}"
+        logger.warning(f"Error to add new car. {error_msg}")
         return {"error": "Car not added"}, 500
 
     return {"message": f"Car added to customer {car.owner_id}."}, 200
 
 
-@car_bp.delete("/cars/<id>", responses={"200": MessageSchema})
+@car_bp.delete("/cars/<id>", responses={"200": MessageSchema}, security=security)
 def remove_car_by_id(path: IdPathSchema):
     """Remove car by the ID from the database"""
+    token = str(request.authorization).split(" ")[-1]
+    token_check = jwt_decode(token)
+
+    if not token_check:
+        error_msg = "Login required for insert operation"
+        return {"error": error_msg}, 403
+
     session = Session()
     car_id = path.id
     response = session.query(Car).filter(Car.id == car_id).delete()
